requirements.txt
google-generativeai
deepgram-sdk
fastapi==0.115.0
uvicorn[standard]==0.32.0
requests==2.31.0
python-dotenv==1.0.0
firebase-admin
google-cloud-firestore
python-multipart
slowapi
neo4j>=5.15.0
